Reorganize bench.rs
